==> Sentiment_Analysis/5_Financial_Metics_Analysis_Part1.ipynb <==
import pandas as pd

# Load finance dataset
df_finance = pd.read_csv("D:\\NIT_Trichy Project\\Sentiment_Analysis\\cleaned_finance_cleaned.csv")

# Quick preview
print(df_finance.columns)
print(df_finance.head())

# Convert relevant time columns to datetime format
df_finance['createdAt'] = pd.to_datetime(df_finance['createdAt'], errors='coerce')
df_finance['bonding_completeTime'] = pd.to_datetime(df_finance['bonding_completeTime'], errors='coerce')

market_entry_time = df_finance['createdAt'].min()
print("Market Entry Time:", market_entry_time)

# Convert mCap to numeric (in case it’s string or contains commas)
df_finance['mCap'] = pd.to_numeric(df_finance['mCap'], errors='coerce')

# Find the row with the highest market cap
peak_row = df_finance.loc[df_finance['mCap'].idxmax()]
peak_mcap = peak_row['mCap']
peak_token = peak_row['token_name']
peak_time = peak_row['createdAt']

print("Peak Market Capitalization:")
print(f"Token: {peak_token}")
print(f"Value: {peak_mcap}")
print(f"Time: {peak_time}")


missing_mcap_count = df_finance['mCap'].isna().sum()
print("Number of missing values in 'mCap' column:", missing_mcap_count)
